requirements/testing.txt
# SHA1:623feb0dd2b6bd376238ecf75069bc82136c2d70
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r development.txt
-r integration.txt
-e file:.
    # via
    #   -r requirements/base.in
    #   -r requirements/development.in
    #   -r requirements/testing.in
astroid==2.6.6
    # via pylint
cachetools==5.2.0
    # via google-auth
coverage==5.5
    # via pytest-cov
db-dtypes==1.0.5
    # via pandas-gbq
docker==5.0.0
    # via -r requirements/testing.in
flask-testing==0.8.1
    # via -r requirements/testing.in
freezegun==1.1.0
    # via -r requirements/testing.in
google-api-core[grpc]==2.11.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   pandas-gbq
    #   sqlalchemy-bigquery
google-auth==2.14.1
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   pandas-gbq
    #   pydata-google-auth
    #   sqlalchemy-bigquery
google-auth-oauthlib==0.7.1
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery[bqstorage,pandas]==3.4.0
    # via
    #   apache-bridge
    #   pandas-gbq
    #   sqlalchemy-bigquery
google-cloud-bigquery-storage==2.16.2
    # via
    #   google-cloud-bigquery
    #   pandas-gbq
    #   sqlalchemy-bigquery
google-cloud-core==2.3.2
    # via google-cloud-bigquery
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-bigquery
googleapis-common-protos==1.57.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.51.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.51.1
    # via google-api-core
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
oauthlib==3.2.2
    # via requests-oauthlib
openapi-schema-validator==0.1.5
    # via openapi-spec-validator
openapi-spec-validator==0.3.1
    # via -r requirements/testing.in
pandas-gbq==0.18.1
    # via apache-bridge
parameterized==0.8.1
    # via -r requirements/testing.in
proto-plus==1.22.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
protobuf==4.21.10
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydata-google-auth==1.4.0
    # via pandas-gbq
pyfakefs==4.5.6
    # via -r requirements/testing.in
pyhive[presto]==0.6.5
    # via apache-bridge
pylint==2.9.6
    # via -r requirements/testing.in
pytest==6.2.4
    # via
    #   -r requirements/testing.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==2.12.1
    # via -r requirements/testing.in
pytest-mock==3.6.1
    # via -r requirements/testing.in
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sqlalchemy-bigquery==1.5.0
    # via apache-bridge
statsd==3.3.0
    # via -r requirements/testing.in
trino==0.319.0
    # via apache-bridge
websocket-client==1.2.0
    # via docker
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
